+++ rust/message-backup/src/scramble.rs
@@ -599,8 +599,9 @@ impl Visit<Scrambler> for proto::group::GroupSnapshot {
             membersPendingProfileKey,
             membersPendingAdminApproval,
             inviteLinkPassword,
-            announcements_only: _,
-            members_banned,
+            announcementsOnly: _,
+            membersBanned,
+            terminated: _,
             special_fields: _,
         } = self;
         title.accept(visitor);
@@ -614,7 +615,7 @@ impl Visit<Scrambler> for proto::group::GroupSnapshot {
         membersPendingProfileKey.accept(visitor);
         membersPendingAdminApproval.accept(visitor);
         inviteLinkPassword.randomize(&mut visitor.rng);
-        members_banned.accept(visitor);
+        membersBanned.accept(visitor);
     }
 }
 
@@ -652,19 +653,19 @@ impl Visit<Scrambler> for proto::group::AccessControl {
 impl Visit<Scrambler> for proto::group::Member {
     fn accept(&mut self, visitor: &mut Scrambler) {
         let Self {
-            user_id,
+            userId,
             role: _,
             joinedAtVersion: _,
-            label_emoji,
-            label_string,
+            labelEmoji,
+            labelString,
             special_fields: _,
         } = self;
-        visitor.replace_service_id(user_id);
-        if !label_emoji.is_empty() {
-            *label_emoji = REPLACEMENT_EMOJI.to_string();
+        visitor.replace_service_id(userId);
+        if !labelEmoji.is_empty() {
+            *labelEmoji = REPLACEMENT_EMOJI.to_string();
         }
-        if !label_string.is_empty() {
-            label_string.randomize(&mut visitor.rng);
+        if !labelString.is_empty() {
+            labelString.randomize(&mut visitor.rng);
         }
     }
 }
@@ -685,22 +686,22 @@ impl Visit<Scrambler> for proto::group::MemberPendingProfileKey {
 impl Visit<Scrambler> for proto::group::MemberPendingAdminApproval {
     fn accept(&mut self, visitor: &mut Scrambler) {
         let Self {
-            user_id,
+            userId,
             timestamp: _,
             special_fields: _,
         } = self;
-        visitor.replace_service_id(user_id);
+        visitor.replace_service_id(userId);
     }
 }
 
 impl Visit<Scrambler> for proto::group::MemberBanned {
     fn accept(&mut self, visitor: &mut Scrambler) {
         let Self {
-            user_id,
+            userId,
             timestamp: _,
             special_fields: _,
         } = self;
-        visitor.replace_service_id(user_id);
+        visitor.replace_service_id(userId);
     }
 }
 
@@ -1340,6 +1341,7 @@ impl Visit<Scrambler> for proto::group_change_chat_update::Update {
                 Update::GroupV2MigrationDroppedMembersUpdate(update) => update.accept(visitor),
                 Update::GroupSequenceOfRequestsAndCancelsUpdate(update) => update.accept(visitor),
                 Update::GroupExpirationTimerUpdate(update) => update.accept(visitor),
+                Update::GroupTerminateChangeUpdate(update) => update.accept(visitor),
             }
         }
     }
@@ -1781,6 +1783,18 @@ impl Visit<Scrambler> for proto::GroupExpirationTimerUpdate {
     fn accept(&mut self, visitor: &mut Scrambler) {
         let Self {
             expiresInMs: _,
+            updaterAci,
+            special_fields: _,
+        } = self;
+        if let Some(updater) = updaterAci {
+            visitor.replace_service_id(updater);
+        }
+    }
+}
+
+impl Visit<Scrambler> for proto::GroupTerminateChangeUpdate {
+    fn accept(&mut self, visitor: &mut Scrambler) {
+        let Self {
             updaterAci,
             special_fields: _,
         } = self;